Add Poitiers

+++ src/main/java/com/instantsystem/recruitment/carparkcontext/readside/domain/model/CarPark.java
@@ -3,7 +3,7 @@ package com.instantsystem.recruitment.carparkcontext.readside.domain.model;
 import lombok.Builder;
 
 @Builder
-public record CarPark(String name, int distance,
+public record CarPark(String name, double distance,
                       Coordinates coordinates, Capacity capacity) {
 
 }
